refactor returnPhoneObject for 995 and 996 into shared utils helpers
