+++ src/applications/appeals/995/tests/utils/contactInfo.unit.spec.js
@@ -8,6 +8,8 @@ import {
   hasHomeAndMobilePhone,
 } from '../../utils/contactInfo';
 
+import { returnPhoneObject } from '../../../shared/utils/helpers';
+
 const getPhone = ({
   country = '1',
   area = '800',
