+++ src/applications/appeals/996/utils/helpers.js
@@ -95,23 +95,3 @@ export const getItemSchema = (schema, index) => {
  */
 export const calculateIndexOffset = (index, contestableIssuesLength) =>
   index - contestableIssuesLength;
-
-/**
- * Return a phone number object
- * @param {String} phone - phone number string to convert to an object
- * @return {phoneObject}
- */
-export const returnPhoneObject = phone => {
-  const result = {
-    countryCode: '',
-    areaCode: '',
-    phoneNumber: '',
-    phoneNumberExt: '',
-  };
-  if (typeof phone === 'string' && phone?.length === 10) {
-    result.countryCode = '1';
-    result.areaCode = phone.slice(0, 3);
-    result.phoneNumber = phone.slice(-7);
-  }
-  return result;
-};
