+++ src/applications/appeals/shared/utils/helpers.js
@@ -11,3 +11,23 @@ export const getItemSchema = (schema, index) => {
   }
   return itemSchema.additionalItems;
 };
+
+/**
+ * Return a phone number object
+ * @param {String} phone - phone number string to convert to an object
+ * @return {phoneObject}
+ */
+export const returnPhoneObject = phone => {
+  const result = {
+    countryCode: '',
+    areaCode: '',
+    phoneNumber: '',
+    phoneNumberExt: '',
+  };
+  if (typeof phone === 'string' && phone?.length === 10) {
+    result.countryCode = '1';
+    result.areaCode = phone.slice(0, 3);
+    result.phoneNumber = phone.slice(-7);
+  }
+  return result;
+};
